+++ vct/__init__.py
@@ -4,75 +4,10 @@ from sqlalchemy.orm import sessionmaker, session
 from sqlalchemy import create_engine
 
 from .databases import Tournament, Map, Agent, Comp, Team, Match, Referall
-from .functions import data_check, choice_check, int_input
+from .functions import data_check, choice_check, int_input, setup
 from .new_game import new_game
 from .viewer import data_viewer
 from .get_data import VLRScrape
-
-
-# creates the initial maps, agents and teams tables
-def setup(tournament: Tournament, session: session.Session):
-    """Function to create all the required fields in the databases for a new tournament.
-
-    Parameter
-    ---------
-    tournament : Tournament
-        The database the contains base information of the tournament to be created.
-    session : session.Session"""
-
-    maps = tournament.map_pool.split(" - ")
-    agents = tournament.agent_pool.split(" - ")
-    teams = tournament.team_pool.split(" - ")
-    tournament = tournament.tournament
-
-    ovr_map = Map(tournament=tournament,
-                  map="Overall",
-                  games=0,
-                  ct_wins=0,
-                  t_wins=0)
-    session.add(ovr_map)
-
-    for agent in agents:
-        ovr_agent = Agent(tournament=tournament,
-                          map="Overall",
-                          agent=agent,
-                          games=0,
-                          wins=0)
-        session.add(ovr_agent)
-
-    for team in teams:
-        ovr_team = Team(tournament=tournament,
-                        map="Overall",
-                        team=team,
-                        games=0,
-                        wins=0)
-        session.add(ovr_team)
-
-    for map in maps:
-        map_row = Map(tournament=tournament,
-                      map=map,
-                      games=0,
-                      ct_wins=0,
-                      t_wins=0)
-        session.add(map_row)
-
-        for agent in agents:
-            agent_row = Agent(tournament=tournament,
-                              map=map,
-                              agent=agent,
-                              games=0,
-                              wins=0)
-            session.add(agent_row)
-
-        for team in teams:
-            team_row = Team(tournament=tournament,
-                            map=map,
-                            team=team,
-                            games=0,
-                            wins=0)
-            session.add(team_row)
-
-    session.commit()
 
 
 # reloads all match data
@@ -359,13 +294,15 @@ def vlr_scraper(session):
     while True:
         url = input("Enter the URL to be scraped")
         split_url = url.split("/")
-        if split_url[2] == "www.vlr.gg":
+        if len(split_url) < 4:
+            print("INVALID URL: Must be a www.vlr.gg match or event url")
+        elif split_url[2] == "www.vlr.gg":
             if split_url[3] == "event":
                 scraper.add_tournaments(url)
             else:
                 scraper.add_matches(url)
         else:
-            print("INVALID URL: Must be a www.vlr.gg url")
+            print("INVALID URL: Must be a www.vlr.gg match or event url")
 
         done = choice_check("Would you like to do anything else? (y/n) ",
                             ["y", "n"])
